Support configuration by monitor id only
This allows you to specify a monitor id alone for both output matching
and for output configuration.

I've got a dock that every time I plug into it it show up as a
different output (e.g. dp-4 then dp-5). This made it difficult to get
hot plugging to work properly.

The change allows for the output configuration to use the monitor id
instead of the output name.

I also changed randr Connected() method to be able to match on just
the output name. This isn't strictly necessary since you can use
wildcards but it seemed a bit confusing to require a wildcard for the
outputs_connected config and then a strict match for the configure_row
config.